rename stream to timeline
