+++ app/src/pages/index.ts
@@ -2,7 +2,7 @@ export * from './Associations'
 export * from './Explorer'
 export * from './Notifications'
 export * from './Settings'
-export * from './StreamPage'
+export * from './TimelinePage'
 export * from './Entity'
 export * from './MessagePage'
 export * from './Devtool'
